requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
faiss-cpu==1.7.4
openpyxl==3.1.5
python-dotenv==1.0.0
google-generativeai==0.3.2
